refactor: move directory process types
