+++ packages/cli/src/index.ts
@@ -13,6 +13,8 @@ export {
     ResolveRequests,
     typedConfiguration,
 } from './types';
-export { DirectoryProcessService } from './directory-process-service/directory-process-service';
-export type { DirectoryProcessServiceOptions } from './directory-process-service/types';
+export {
+    DirectoryProcessService,
+    DirectoryProcessServiceOptions,
+} from './directory-process-service/directory-process-service';
 export type { CodeMod } from './code-mods/types';
